bioprocess/requirements.txt
aniso8601==3.0.2
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==1.0.2
Flask-Cors==3.0.6
Flask-JWT==0.3.2
Flask-JWT-Simple==0.0.3
Flask-RESTful==0.3.6
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
PyJWT==1.6.4
pymongo==3.6.1
pytz==2018.4
requests==2.19.1
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
Cerberus==1.1
pandas